Program.cs: Remove DefaultValue for counter/length.
Introducing -w|--words means there's a separate default length for derived passwords and passphrases. Then displaying the default for the counter makes it look like there's no default for the length.

This also fixes an issue where you could specify --length 20 with --words and it wouldn't produce an error despite the maximum being 12 words. Instead, it would default to 8 words because 20 was the default length value and hence valid.