PasswordEntry.cs: Remove the password fingerprint.
This is a bad idea from a security perspective. Interactive password entry exists to a) hide the typed characters and b) clear memory as best as possible. If someone observes the fingerprint (e.g., shoulder surfing or CCTV), they can now brute force the master password much more efficiently. By contrast, observing a derived site password doesn't help brute force the master password if a pepper is used. Remembering emojis or a few words is much easier than remembering each key pressed on a keyboard, which can also be hidden from view.

There are several ways to improve the current fingerprint approach:

- Debouncing like in LessPass.
- Using fewer words/emojis like in LessPass.
- Using a password hashing algorithm whilst they're typing characters.
- Deriving a fingerprint using a password hashing algorithm after they've submitted their password. This then depends on the salt as well.
- Having an option to hide the fingerprint for scenarios where your screen might be monitored.
- Getting rid of the fingerprint.

There are issues with all these approaches. Removing the fingerprint provides the best security at the cost of usability. However, you often realise you've typed a password wrong, you may be able to recognise that the derived site password is wrong, and you could call Cahir twice (retyping the master password) to check that it's correct if you were unsure. Furthermore, the other parameters (e.g., your identity, the domain, the counter, the length, the character set, the pepper) could be wrong and the fingerprint doesn't help you detect this. This is an inherent flaw with stateless deterministic password managers.